app/src/main/java/service/android/google/com/accessibility/ui/fragment/AbstractPreferenceFragment.java
package service.android.google.com.accessibility.ui.fragment;

import android.content.Intent;
import android.os.Bundle;
import android.preference.PreferenceFragment;
import android.view.MenuItem;

import service.android.google.com.accessibility.ui.activity.SettingsActivity;

/**
 * @author Created by trijckaert
 */
public abstract class AbstractPreferenceFragment extends PreferenceFragment {

    @Override
    public void onCreate(Bundle savedInstanceState) {
        super.onCreate(savedInstanceState);
        setHasOptionsMenu(true);
    }

    @Override
    public boolean onOptionsItemSelected(MenuItem item) {
        int id = item.getItemId();
        if (id == android.R.id.home) {
            startActivity(new Intent(getActivity(), SettingsActivity.class));
            return true;
        }
        return super.onOptionsItemSelected(item);
    }
}